+++ packages/lib/passport-interface/src/genericIssuanceTypes.ts
@@ -239,7 +239,6 @@ export type PretixPipelineDefinition = z.infer<
 >;
 
 export enum CSVPipelineOutputType {
-  RSAImage = "RSAImage",
   EdDSAMessage = "EdDSAMessage",
   EdDSATicket = "EdDSATicket"
 }
