+++ packages/pcd/eddsa-message-pcd/package.json
@@ -28,6 +28,7 @@
   "dependencies": {
     "@pcd/eddsa-pcd": "0.5.0",
     "@pcd/pcd-types": "0.10.0",
+    "@pcd/semaphore-signature-pcd": "0.10.0",
     "@trufflesuite/bigint-buffer": "^1.1.10",
     "chai": "^4.3.7",
     "json-bigint": "^1.0.0",
